+++ actionpack/test/controller/render_xml_test.rb
@@ -81,7 +81,7 @@ class RenderXmlTest < ActionController::TestCase
   end
 
   def test_should_render_formatted_xml_erb_template
-    get :formatted_xml_erb, :format => :xml
+    get :formatted_xml_erb, format: :xml
     assert_equal '<test>passed formatted xml erb</test>', @response.body
   end
 
@@ -91,7 +91,7 @@ class RenderXmlTest < ActionController::TestCase
   end
 
   def test_should_use_implicit_content_type
-    get :implicit_content_type, :format => 'atom'
+    get :implicit_content_type, format: 'atom'
     assert_equal Mime::ATOM, @response.content_type
   end
 end